+++ plniconweb/src/components/Checkbox.tsx
@@ -21,7 +21,7 @@ function Checkbox({
     <>
       <div className="flex h-auto w-full justify-start lg:justify-center gap-3">
         <input
-          className={`before:w-4 before:h-3 checked:before:bg-[url('./assets/check_icon.svg')] before:inline-block before:bg-center before:bg-contain before:bg-no-repeat before:absolute before:-top-[-3px] relative h-[20px] w-[20px] shrink-0 cursor-pointer appearance-none rounded border-2 border-slate-500 outline-none before:border-white before:text-center before:font-bold before:text-white checked:border-blue-graph checked:bg-blue-graph hover:border-blue-hover checked:hover:border-blue-hover checked:hover:bg-blue-hover focus:outline focus:outline-blue-graph checked:focus:border-blue-graph focus:hover:border-blue-hover focus:hover:outline-blue-hover active:border-blue-click checked:active:border-blue-click checked:active:bg-blue-click focus:active:border-blue-click focus:active:outline-blue-click disabled:opacity-30 disabled:hover:border-slate-500`}
+          className={`before:w-4 before:h-3 checked:before:bg-[url('/assets/check_icon.svg')] before:inline-block before:bg-center before:bg-contain before:bg-no-repeat before:absolute before:-top-[-3px] relative h-[20px] w-[20px] shrink-0 cursor-pointer appearance-none rounded border-2 border-slate-500 outline-none before:border-white before:text-center before:font-bold before:text-white checked:border-blue-graph checked:bg-blue-graph hover:border-blue-hover checked:hover:border-blue-hover checked:hover:bg-blue-hover focus:outline focus:outline-blue-graph checked:focus:border-blue-graph focus:hover:border-blue-hover focus:hover:outline-blue-hover active:border-blue-click checked:active:border-blue-click checked:active:bg-blue-click focus:active:border-blue-click focus:active:outline-blue-click disabled:opacity-30 disabled:hover:border-slate-500`}
           type="checkbox"
           id={id}
           checked={checked}
