+++ plniconweb/src/pages/other_pages/coming_soon.tsx
@@ -4,7 +4,7 @@ function ComingSoon() {
     <>
       <Navbar/>
       <div className="w-full h-screen">
-        <div className="w-full h-full bg-[url(./assets/coming_soon_alt.svg)] bg-cover bg-no-repeat"></div>
+        <div className="w-full h-full bg-[url(/assets/coming_soon_alt.svg)] bg-cover bg-no-repeat"></div>
       </div>
     </>
   );
